Separate reconfiguration test Test parameters are too long to be in engine test together
